changed float to double

+++ BadilloCallender_Seq_1024.cpp
@@ -37,8 +37,8 @@ void lu_decomposition(const Matrix_t &matrix, Matrix_t &lower_matrix, Matrix_t &
 
 /// @brief Computes the determinant of a triangular matrix.
 /// @param matrix Triangular matrix.
-/// @return The determinant, a float.
-float determinant_triangular(const Matrix_t &matrix);
+/// @return The determinant, a double.
+double determinant_triangular(const Matrix_t &matrix);
 
 /// @brief Creates an NxN identity matrix.
 /// @param size N, size of identity matrix.
@@ -59,7 +59,7 @@ void print_matrix(const Matrix_t &matrix);
 int main()
 {
     // Determinants for original matrix, lower triangular matrix, and upper triangular matrix
-    float determinant, l_det, u_det;
+    double determinant, l_det, u_det;
 
     // Matrix to decompose
     Matrix_t matrix = random_matrix(matrix_size, -1.0, 1.0);
@@ -171,12 +171,12 @@ void lu_decomposition(const Matrix_t &matrix, Matrix_t &lower_matrix, Matrix_t &
     }
 }
 
-float determinant_triangular(const Matrix_t &matrix)
+double determinant_triangular(const Matrix_t &matrix)
 {
     size_t size = matrix.size();
 
     // 0x0 matrix has determinant of 1
-    float determinant = 1;
+    double determinant = 1;
 
     // Compute determinant of triangular matrix
     for (size_t i = 0; i < size; i++)
